[c++] Fix some MSVC static analysis issues
* `isspace` takes ints, not chars, so promote.
* Instead of dealing with a potentially non-null terminated string when
  finding a field by ordinal in Simple JSON, attempt to parse the field
  ID and then compare the numeric values.
* Indicate that the `die` function never returns so that the analyzer
  can tell that we've checked for things like malloc returning null.
* The printf format string for unsigned ints is u, not d.
* Disable the code analysis warning C6326, "constant constant
  comparison". A lot of our template functions trigger this when they do
  things like: if (T == traits<U>::some_value) { ... } else { ... }

+++ cpp/inc/bond/core/detail/cmdargs.h
@@ -535,10 +535,10 @@ protected:
 
         for (int begin = 0; i < static_cast<int>(formated.length()); begin += 80, i = begin + 79 - indent)
         {
-            while (i >= begin && !isspace(formated[i]))
+            while (i >= begin && !isspace(static_cast<int>(formated[i])))
                 --i;
 
-            while (i < static_cast<int>(formated.length()) && isspace(formated[i]))
+            while (i < static_cast<int>(formated.length()) && isspace(static_cast<int>(formated[i])))
                 ++i;
 
             formated.insert(i, begin + 80 - i, ' ');
